+++ src/js/gutenberg-overrides/components/index.js
@@ -1,6 +1,2 @@
-import def from 'gutenberg/components/index?source=node_modules';
-
 export { default as CategorySelect } from './query-controls/category-select.js';
-
-export default def;
 export * from 'gutenberg/components/index?source=node_modules';